streamlit_newsCollection/dev/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.25
botocore==1.35.25
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
DateTime==5.5
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.2
grpcio-tools==1.63.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.8.2
networkx==3.2.1
numpy==2.1.1
openai==0.28.0
packaging==24.1
pandas==2.2.3
param==2.1.1
pillow==10.4.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.5
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stability-sdk==0.8.6
streamlit==1.38.0
sympy==1.12
tenacity==8.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.12.1
zope.interface==7.0.3
